move any library packages into common
