+++ internal/common/ingest/ingestion_pipeline_test.go
@@ -2,6 +2,7 @@ package ingest
 
 import (
 	"context"
+	"github.com/G-Research/armada/internal/common/pulsarutils"
 	"testing"
 	"time"
 
@@ -12,7 +13,6 @@ import (
 
 	"github.com/G-Research/armada/internal/armada/configuration"
 	"github.com/G-Research/armada/internal/common/ingest/metrics"
-	"github.com/G-Research/armada/internal/pulsarutils"
 	"github.com/G-Research/armada/pkg/armadaevents"
 )
