+++ internal/common/pulsarutils/errors.go
