+++ internal/common/pulsarutils/eventsequence.go
