+++ internal/lookoutingesterv2/instructions/instructions_test.go
@@ -2,6 +2,7 @@ package instructions
 
 import (
 	"context"
+	"github.com/G-Research/armada/internal/common/pulsarutils"
 	"strings"
 	"testing"
 
@@ -19,7 +20,6 @@ import (
 	"github.com/G-Research/armada/internal/common/ingest/testfixtures"
 	"github.com/G-Research/armada/internal/lookoutingesterv2/metrics"
 	"github.com/G-Research/armada/internal/lookoutingesterv2/model"
-	"github.com/G-Research/armada/internal/pulsarutils"
 	"github.com/G-Research/armada/pkg/armadaevents"
 )
